+++ ui/src/main/kotlin/com/processout/sdk/ui/checkout/PODynamicCheckoutLauncher.kt
@@ -112,7 +112,7 @@ class PODynamicCheckoutLauncher private constructor(
     }
 
     private fun dispatchInvoice() {
-        eventDispatcher.subscribeForRequest<PODynamicCheckoutInvoiceRequest>(
+        eventDispatcher.subscribeForRequest<DynamicCheckoutInvoiceRequest>(
             coroutineScope = scope
         ) { request ->
             scope.launch {
