--- ui/src/main/kotlin/com/processout/sdk/ui/checkout/delegate/PODynamicCheckoutInvoiceRequest.kt
@@ -1,14 +1,0 @@
-package com.processout.sdk.ui.checkout.delegate
-
-import com.processout.sdk.api.dispatcher.POEventDispatcher
-import com.processout.sdk.api.model.response.POInvoice
-import com.processout.sdk.core.annotation.ProcessOutInternalApi
-import java.util.UUID
-
-/** @suppress */
-@ProcessOutInternalApi
-data class PODynamicCheckoutInvoiceRequest(
-    override val uuid: UUID = UUID.randomUUID(),
-    val currentInvoice: POInvoice,
-    val invalidationReason: PODynamicCheckoutInvoiceInvalidationReason
-) : POEventDispatcher.Request
